features =)

+++ src/helper/channels.h
@@ -157,7 +157,7 @@ bool CHANNELS_LoadBuf();
 void CHANNELS_Next(bool next);
 void CHANNELS_Delete(int16_t i);
 bool CHANNELS_Existing(int16_t i);
-uint8_t CHANNELS_Scanlists(int16_t i);
+uint16_t CHANNELS_Scanlists(int16_t i);
 void CHANNELS_LoadScanlist(CHTypeFilter type, uint16_t n);
 void CHANNELS_LoadBlacklistToLoot();
